requirements.txt
asn1crypto==0.24.0
cachetools==3.1.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
defusedxml==0.6.0
Django==2.2.6
django-allauth==0.40.0
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
httplib2==0.14.0
idna==2.8
oauthlib==3.1.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
pyOpenSSL==19.0.0
PySocks==1.7.1
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.24.2
